Add `React#renderToString` parity checking that can be used anywhere.
